# backend/src/main/java/com/paydue/api/dto/LoginRequest.java
package com.paydue.api.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank @Email String email,
        @NotBlank String password
) {
}