# backend/src/main/java/com/paydue/api/dto/RegisterRequest.java
package com.paydue.api.dto;

import com.paydue.domain.UserRole;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @NotBlank @Email String email,
        @NotBlank @Size(min = 8, max = 72) String password,
        @NotNull UserRole role,
        String name
) {
}