# backend/src/main/java/com/paydue/api/dto/SupplierResponse.java
package com.paydue.api.dto;

import java.time.Instant;

public record SupplierResponse(
        Long id,
        String name,
        String email,
        String gstin,
        boolean msme,
        Instant registeredAt,
        long productCount
) {
}