# openapi.yaml
openapi: 3.0.3
info:
  title: PayDue API
  description: MSME overdue-invoice tracker (supplier-first v1)
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /api/auth/register:
    post:
      summary: Register BUYER, SELLER, or ADMIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already registered
  /api/auth/login:
    post:
      summary: Login and get a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid email or password
  /api/auth/me:
    get:
      summary: Current user from Bearer token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Missing or invalid token
  /api/suppliers:
    get:
      summary: List all suppliers (admin)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupplierResponse'
    post:
      summary: Create supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
  /api/admin/users:
    get:
      summary: List registered users (admin)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummaryResponse'

  /api/suppliers/{supplierId}:
    get:
      summary: Get supplier profile (productCount is computed)
      parameters:
        - $ref: '#/components/parameters/SupplierId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
        '404':
          description: Not found

  /api/suppliers/{supplierId}/products:
    get:
      summary: List SKUs for this supplier
      parameters:
        - $ref: '#/components/parameters/SupplierId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
    post:
      summary: Add SKU
      parameters:
        - $ref: '#/components/parameters/SupplierId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '409':
          description: Duplicate skuCode for this supplier

  /api/buyers:
    post:
      summary: Create a buyer in the global list (does not link yet)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBuyerRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerResponse'
  /api/buyers/{buyerId}:
    get:
      summary: Get buyer profile
      parameters:
        - name: buyerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerResponse'
  /api/buyers/{buyerId}/invoices:
    get:
      summary: Invoices billed to this buyer, with totals and aging
      parameters:
        - name: buyerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceResponse'

  /api/suppliers/{supplierId}/buyers/{buyerId}:
    post:
      summary: Link an existing buyer to this supplier (my customers)
      parameters:
        - $ref: '#/components/parameters/SupplierId'
        - name: buyerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerResponse'
        '409':
          description: Already linked

  /api/suppliers/{supplierId}/buyers:
    get:
      summary: List buyers linked to this supplier only
      parameters:
        - $ref: '#/components/parameters/SupplierId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuyerResponse'

  /api/invoices:
    post:
      summary: Create invoice with line items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '409':
          description: Duplicate invoice number, or buyer not linked to supplier

  /api/suppliers/{supplierId}/invoices:
    get:
      summary: Invoices issued by this supplier, with totals and aging
      parameters:
        - $ref: '#/components/parameters/SupplierId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceResponse'

components:
  parameters:
    SupplierId:
      name: supplierId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    CreateSupplierRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: string }
        gstin: { type: string }
        msme: { type: boolean, default: true }
    SupplierResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string }
        gstin: { type: string }
        msme: { type: boolean }
        registeredAt: { type: string, format: date-time }
        productCount: { type: integer }
    CreateProductRequest:
      type: object
      required: [skuCode, name]
      properties:
        skuCode: { type: string }
        name: { type: string }
        productType: { type: string, default: OTHER }
        shelfLifeDays: { type: integer }
    ProductResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        supplierId: { type: integer, format: int64 }
        skuCode: { type: string }
        name: { type: string }
        productType: { type: string }
        shelfLifeDays: { type: integer }
        active: { type: boolean }
        registeredAt: { type: string, format: date-time }
    CreateBuyerRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: string }
        gstin: { type: string }
    BuyerResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string }
        gstin: { type: string }
        registeredAt: { type: string, format: date-time }
    CreateInvoiceLineRequest:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId: { type: integer, format: int64 }
        quantity: { type: number }
        unitPrice: { type: number }
    CreateInvoiceRequest:
      type: object
      required: [supplierId, buyerId, invoiceNumber, invoiceDate, lines]
      properties:
        supplierId: { type: integer, format: int64 }
        buyerId: { type: integer, format: int64 }
        invoiceNumber: { type: string }
        invoiceDate: { type: string, format: date }
        dueDate: { type: string, format: date, description: Optional; defaults to invoiceDate + 45 days }
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateInvoiceLineRequest'
    InvoiceLineResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        productId: { type: integer, format: int64 }
        skuCode: { type: string }
        description: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        lineTotal: { type: number }
    InvoiceResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        supplierId: { type: integer, format: int64 }
        supplierName: { type: string }
        buyerId: { type: integer, format: int64 }
        buyerName: { type: string }
        invoiceNumber: { type: string }
        invoiceDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        status: { type: string }
        totalAmount: { type: number }
        daysOverdue: { type: integer }
        lines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineResponse'
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        role: { type: string, enum: [BUYER, SELLER, ADMIN] }
        name: { type: string, description: Company name for BUYER or SELLER }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        tokenType: { type: string, example: Bearer }
        userId: { type: integer, format: int64 }
        email: { type: string }
        role: { type: string, enum: [BUYER, SELLER, ADMIN] }
        supplierId: { type: integer, format: int64, nullable: true }
        buyerId: { type: integer, format: int64, nullable: true }
    UserSummaryResponse:
      type: object
      properties:
        userId: { type: integer, format: int64 }
        email: { type: string }
        role: { type: string, enum: [BUYER, SELLER, ADMIN] }
        supplierId: { type: integer, format: int64, nullable: true }
        buyerId: { type: integer, format: int64, nullable: true }
        companyName: { type: string, nullable: true }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT